replace spectest executor
